# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Book Notes Organizer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books/{bookId}/notes:
    get:
      operationId: getNotes
      tags: [notes]
      summary: List all notes for a book
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note for a book chapter
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/notes/{noteId}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Note not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Note not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Note deleted
        "404":
          description: Note not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/quotes:
    get:
      operationId: getQuotes
      tags: [quotes]
      summary: List all quotes for a book
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a highlighted quote
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/quotes/{quoteId}:
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update a quote
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: quoteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteRequest"
      responses:
        "200":
          description: Quote updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: quoteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Quote deleted
        "404":
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Note:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        chapter:
          type: ["string", "null"]
        chapterNumber:
          type: ["integer", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        editingBy:
          type: ["string", "null"]
      required:
        - id
        - bookId
        - userId
        - chapter
        - chapterNumber
        - content
        - createdAt
        - updatedAt
        - editingBy
    CreateNoteRequest:
      type: object
      properties:
        chapter:
          type: string
        chapterNumber:
          type: integer
        content:
          type: string
          minLength: 1
      required:
        - content
    UpdateNoteRequest:
      type: object
      properties:
        chapter:
          type: string
        chapterNumber:
          type: integer
        content:
          type: string
    Quote:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        text:
          type: string
        page:
          type: ["integer", "null"]
        chapter:
          type: ["string", "null"]
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookId
        - userId
        - text
        - page
        - chapter
        - color
        - createdAt
    CreateQuoteRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
        page:
          type: integer
        chapter:
          type: string
        color:
          type: string
      required:
        - text
    UpdateQuoteRequest:
      type: object
      properties:
        text:
          type: string
        page:
          type: integer
        chapter:
          type: string
        color:
          type: string
...