# lib/api-zod/src/generated/types/createQuoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Notes Organizer API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuoteRequest {
  /** @minLength 1 */
  text: string;
  page?: number;
  chapter?: string;
  color?: string;
}