# lib/api-zod/src/generated/types/updateNoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Book Notes Organizer API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateNoteRequest {
  chapter?: string;
  chapterNumber?: number;
  content?: string;
}